+++ syntax/ast.mjs
@@ -112,7 +112,7 @@ export class TermReference extends Expression {
         this.type = "TermReference";
         this.id = id;
         this.attribute = attribute;
-        this.args = args;
+        this.arguments = args;
     }
 }
 
@@ -129,7 +129,7 @@ export class FunctionReference extends Expression {
         super();
         this.type = "FunctionReference";
         this.id = id;
-        this.args = args;
+        this.arguments = args;
     }
 }
 
@@ -242,7 +242,7 @@ export class Annotation extends SyntaxNode {
         super();
         this.type = "Annotation";
         this.code = code;
-        this.args = args;
+        this.arguments = args;
         this.message = message;
     }
 }